change nthpercentile value to u8 - remove cov invalid cov test implicit widening - add comment to pest file about winop names

+++ crates/rtlola-frontend/src/mir.rs
@@ -545,7 +545,7 @@ pub enum WindowOperation {
     /// Aggregation function to return the standard deviation of all values, assumes equal probability.
     StandardDeviation,
     /// Aggregation function to return the Nth-Percentile
-    NthPercentile(usize),
+    NthPercentile(u8),
 }
 
 ////////// Implementations //////////
